requirements.txt
flask
onnxruntime
numpy
pillow
requests
